update delete function

+++ movies/views.py
@@ -41,8 +41,10 @@ def detail(request, pk):
 
 def delete(request, pk):
     movie = Movie.objects.get(pk=pk)
-    movie.delete()
-    return redirect('movies:index')
+    if request.user == movie.user:
+        movie.delete()
+        return redirect('movies:index')
+    return redirect('movies:detail', movie.pk)
 
 
 def update(request, pk):
